move server into separate file
